requirements.txt
asgiref==3.7.2
asttokens==2.4.1
crispy-bootstrap5==2023.10
decorator==5.1.1
Django==4.2.7
django-crispy-forms==2.1
django-js-asset==2.2.0
django-mptt==0.15.0
executing==2.0.1
ipython==8.18.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pytils==0.4.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.0
wcwidth==0.2.12
